Fixed issue with state machine
Fixed bug where state machine code would not compile as a result of variables being initialised within the switch case scope. Ensure all variables are initialised outside of the switch case.

Currently, no checks are being done on whether the sensor is connected or whether the rocket is ready to launch. Not sure how that will be done.